# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fantasy Helper API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface EspnSyncInput {
  /** ESPN espn_s2 cookie value */
  s2: string;
  /** ESPN SWID cookie value */
  swid: string;
  /**
     * Optional ESPN league ID to sync
     * @nullable
     */
  leagueId?: number | null;
  /**
     * Sport type: football, basketball, baseball, hockey (auto-detected if omitted)
     * @nullable
     */
  sport?: string | null;
}

export interface SyncResult {
  success: boolean;
  message: string;
  leaguesSynced: number;
  playersSynced: number;
  lastSyncAt?: string;
}

export interface League {
  id: number;
  espnLeagueId: string;
  name: string;
  season: number;
  sport: string;
  /** @nullable */
  teamCount?: number | null;
  syncedAt: string;
}

export interface Team {
  id: number;
  leagueId: number;
  espnTeamId: string;
  name: string;
  abbrev: string;
  wins: number;
  losses: number;
  ties?: number;
  /** @nullable */
  pointsFor?: number | null;
  /** @nullable */
  pointsAgainst?: number | null;
  /** @nullable */
  waiversPosition?: number | null;
}

export type EspnPublicStatsStatLine = {[key: string]: number};

export interface EspnPublicStats {
  gamesPlayed: number;
  statLine: EspnPublicStatsStatLine;
  /** @nullable */
  injuryDescription?: string | null;
  /** @nullable */
  recentHeadline?: string | null;
  fetchedAt: string;
}

export interface Player {
  id: number;
  teamId: number;
  espnPlayerId: string;
  fullName: string;
  position: string;
  proTeam: string;
  /** @nullable */
  projectedPoints?: number | null;
  /** @nullable */
  avgPoints?: number | null;
  /** @nullable */
  totalPoints?: number | null;
  /** @nullable */
  injuryStatus?: string | null;
  espnPublicStats?: EspnPublicStats | null;
}

export interface TradeFindInput {
  leagueId: number;
  myTeamId: number;
  /** Player IDs I am offering */
  offeredPlayerIds: number[];
  /**
     * Desired trade fairness: 1=heavily favors me, 100=perfectly equal
     * @minimum 1
     * @maximum 100
     */
  fairness: number;
  /** Optional positions to target (e.g. QB, RB, WR) */
  targetPositions?: string[];
  /**
     * Max number of players to receive in a package
     * @minimum 1
     * @maximum 3
     */
  packageSize?: number;
}

export interface TradePackage {
  targetTeamId: number;
  targetTeamName: string;
  record: string;
  /** Player names/details to request back */
  playersToReceive: string[];
  /** Estimated fairness 0-100 */
  fairnessScore: number;
  reasoning: string;
  /** Send It, Consider, or Skip */
  recommendation: string;
}

export interface TradeFindResult {
  myTeamName: string;
  packages: TradePackage[];
  cached: boolean;
}

export interface TradeEvalInput {
  teamAId: number;
  teamBId: number;
  /** Player IDs Team A is giving up */
  teamAPlayers: number[];
  /** Player IDs Team B is giving up */
  teamBPlayers: number[];
}

export interface TradeEvalResult {
  /** @nullable */
  id?: number | null;
  /** Win score for Team A (0-100) */
  winScoreA: number;
  /** Win score for Team B (0-100) */
  winScoreB: number;
  /** AI-generated analysis of the trade */
  analysis: string;
  /** Accept, Decline, or Neutral */
  recommendation: string;
  /** @nullable */
  teamAName?: string | null;
  /** @nullable */
  teamBName?: string | null;
  cached: boolean;
  /** @nullable */
  evaluatedAt?: string | null;
}

export interface LeagueSyncStatus {
  id: number;
  name: string;
  sport: string;
  autoSyncEnabled: boolean;
  /** @nullable */
  lastAutoSyncAt?: string | null;
  /** @nullable */
  lastAutoSyncError?: string | null;
}

export interface SyncStatus {
  hasCredentials: boolean;
  leagues: LeagueSyncStatus[];
}

export interface SyncRefreshResult {
  success: boolean;
  leaguesSynced: number;
  playersSynced: number;
  errors: string[];
  /** @nullable */
  lastSyncAt?: string | null;
}

export interface DeleteResult {
  success: boolean;
  deleted: number;
}

export interface InsightTip {
  /** trade, waiver, lineup, or general */
  type: string;
  /** high, medium, or low */
  priority: string;
  /** @nullable */
  player?: string | null;
  message: string;
}

export interface TeamInsightsResult {
  teamId: number;
  teamName: string;
  insights: string;
  tips: InsightTip[];
  /** @nullable */
  standingsRank?: number | null;
  /** @nullable */
  totalTeams?: number | null;
  cached: boolean;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface DashboardSummary {
  leagueCount: number;
  teamCount: number;
  playerCount: number;
  tradeEvalCount: number;
  /** @nullable */
  lastSyncAt?: string | null;
  topPlayers?: Player[];
}

export type ListLeaguesParams = {
sport?: string;
};

export type ListPlayersParams = {
leagueId?: number;
position?: string;
};

export type GetDashboardSummaryParams = {
sport?: string;
};