# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fantasy Helper API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: sync
    description: ESPN sync operations
  - name: leagues
    description: League operations
  - name: teams
    description: Team operations
  - name: players
    description: Player operations
  - name: trade
    description: AI Trade Lab operations
  - name: dashboard
    description: Dashboard summary operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /sync/status:
    get:
      operationId: getSyncStatus
      tags: [sync]
      summary: Get auto-sync status for all leagues
      responses:
        "200":
          description: Sync status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncStatus"

  /sync/refresh:
    post:
      operationId: triggerSyncRefresh
      tags: [sync]
      summary: Immediately re-sync all leagues with stored credentials
      responses:
        "200":
          description: Sync completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncRefreshResult"
        "400":
          description: No stored credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /sync-espn:
    post:
      operationId: syncEspn
      tags: [sync]
      summary: Sync ESPN fantasy data
      description: Accepts espn_s2 and SWID tokens, fetches and caches ESPN data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EspnSyncInput"
      responses:
        "200":
          description: Sync successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /leagues:
    get:
      operationId: listLeagues
      tags: [leagues]
      summary: List all leagues
      parameters:
        - name: sport
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of leagues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/League"

  /leagues/{id}/enrich:
    post:
      operationId: enrichLeague
      tags: [leagues]
      summary: Enrich all players in a league with ESPN public stats
      description: Fetches real per-game stats, injury context, and recent news from the public ESPN API for every player in the league. Runs asynchronously — returns immediately.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Enrichment started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: League not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /leagues/{id}:
    get:
      operationId: getLeague
      tags: [leagues]
      summary: Get a league by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: League detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/League"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteLeague
      tags: [leagues]
      summary: Delete a league and all its data
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: League deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /leagues/{leagueId}/teams:
    get:
      operationId: listTeams
      tags: [teams]
      summary: List teams in a league
      parameters:
        - name: leagueId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"

  /teams/{id}:
    get:
      operationId: getTeam
      tags: [teams]
      summary: Get a team by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Team detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /teams/{teamId}/players:
    get:
      operationId: listTeamPlayers
      tags: [players]
      summary: List players on a team
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of players
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Player"

  /players:
    get:
      operationId: listPlayers
      tags: [players]
      summary: List all cached players
      parameters:
        - name: leagueId
          in: query
          required: false
          schema:
            type: integer
        - name: position
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of players
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Player"

  /trade/evaluate:
    post:
      operationId: evaluateTrade
      tags: [trade]
      summary: AI trade evaluator
      description: Evaluates a trade using Groq AI with cached roster context
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeEvalInput"
      responses:
        "200":
          description: Trade evaluation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeEvalResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /trade/find:
    post:
      operationId: findTrades
      tags: [trade]
      summary: AI Trade Finder
      description: Scans all opposing teams and suggests trade packages matching the desired fairness level
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeFindInput"
      responses:
        "200":
          description: Trade finder results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeFindResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /teams/{teamId}/insights:
    get:
      operationId: getTeamInsights
      tags: [teams]
      summary: AI insights and tips for a team
      description: Generates AI-powered roster tips, trade suggestions, and waiver advice based on standings and player data
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: AI insights result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamInsightsResult"
        "404":
          description: Team not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary stats
      description: Returns counts, last sync time, top players, and league overview
      parameters:
        - name: sport
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /dashboard/recent-trades:
    get:
      operationId: listRecentTrades
      tags: [dashboard]
      summary: List recent AI trade evaluations
      responses:
        "200":
          description: Recent trade evaluations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TradeEvalResult"
    delete:
      operationId: clearAllTrades
      tags: [dashboard]
      summary: Delete all trade evaluations
      responses:
        "200":
          description: All trades deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"

  /dashboard/recent-trades/{id}:
    delete:
      operationId: deleteTrade
      tags: [dashboard]
      summary: Delete a single trade evaluation by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trade deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "404":
          description: Trade not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    EspnSyncInput:
      type: object
      required: [s2, swid]
      properties:
        s2:
          type: string
          description: ESPN espn_s2 cookie value
        swid:
          type: string
          description: ESPN SWID cookie value
        leagueId:
          type: ["integer", "null"]
          description: Optional ESPN league ID to sync
        sport:
          type: ["string", "null"]
          description: "Sport type: football, basketball, baseball, hockey (auto-detected if omitted)"

    SyncResult:
      type: object
      required: [success, message, leaguesSynced, playersSynced]
      properties:
        success:
          type: boolean
        message:
          type: string
        leaguesSynced:
          type: integer
        playersSynced:
          type: integer
        lastSyncAt:
          type: string
          format: date-time

    League:
      type: object
      required: [id, espnLeagueId, name, season, sport, syncedAt]
      properties:
        id:
          type: integer
        espnLeagueId:
          type: string
        name:
          type: string
        season:
          type: integer
        sport:
          type: string
        teamCount:
          type: ["integer", "null"]
        syncedAt:
          type: string
          format: date-time

    Team:
      type: object
      required: [id, leagueId, espnTeamId, name, abbrev, wins, losses]
      properties:
        id:
          type: integer
        leagueId:
          type: integer
        espnTeamId:
          type: string
        name:
          type: string
        abbrev:
          type: string
        wins:
          type: integer
        losses:
          type: integer
        ties:
          type: integer
        pointsFor:
          type: ["number", "null"]
        pointsAgainst:
          type: ["number", "null"]
        waiversPosition:
          type: ["integer", "null"]

    EspnPublicStats:
      type: object
      required: [gamesPlayed, statLine, fetchedAt]
      properties:
        gamesPlayed:
          type: integer
        statLine:
          type: object
          additionalProperties:
            type: number
        injuryDescription:
          type: ["string", "null"]
        recentHeadline:
          type: ["string", "null"]
        fetchedAt:
          type: string
          format: date-time

    Player:
      type: object
      required: [id, teamId, espnPlayerId, fullName, position, proTeam]
      properties:
        id:
          type: integer
        teamId:
          type: integer
        espnPlayerId:
          type: string
        fullName:
          type: string
        position:
          type: string
        proTeam:
          type: string
        projectedPoints:
          type: ["number", "null"]
        avgPoints:
          type: ["number", "null"]
        totalPoints:
          type: ["number", "null"]
        injuryStatus:
          type: ["string", "null"]
        espnPublicStats:
          oneOf:
            - $ref: "#/components/schemas/EspnPublicStats"
            - type: "null"

    TradeFindInput:
      type: object
      required: [leagueId, myTeamId, offeredPlayerIds, fairness]
      properties:
        leagueId:
          type: integer
        myTeamId:
          type: integer
        offeredPlayerIds:
          type: array
          items:
            type: integer
          description: Player IDs I am offering
        fairness:
          type: integer
          minimum: 1
          maximum: 100
          description: "Desired trade fairness: 1=heavily favors me, 100=perfectly equal"
        targetPositions:
          type: array
          items:
            type: string
          description: Optional positions to target (e.g. QB, RB, WR)
        packageSize:
          type: integer
          minimum: 1
          maximum: 3
          description: Max number of players to receive in a package

    TradePackage:
      type: object
      required: [targetTeamId, targetTeamName, record, playersToReceive, fairnessScore, reasoning, recommendation]
      properties:
        targetTeamId:
          type: integer
        targetTeamName:
          type: string
        record:
          type: string
        playersToReceive:
          type: array
          items:
            type: string
          description: Player names/details to request back
        fairnessScore:
          type: integer
          description: Estimated fairness 0-100
        reasoning:
          type: string
        recommendation:
          type: string
          description: "Send It, Consider, or Skip"

    TradeFindResult:
      type: object
      required: [packages, myTeamName, cached]
      properties:
        myTeamName:
          type: string
        packages:
          type: array
          items:
            $ref: "#/components/schemas/TradePackage"
        cached:
          type: boolean

    TradeEvalInput:
      type: object
      required: [teamAId, teamBId, teamAPlayers, teamBPlayers]
      properties:
        teamAId:
          type: integer
        teamBId:
          type: integer
        teamAPlayers:
          type: array
          items:
            type: integer
          description: Player IDs Team A is giving up
        teamBPlayers:
          type: array
          items:
            type: integer
          description: Player IDs Team B is giving up

    TradeEvalResult:
      type: object
      required: [winScoreA, winScoreB, analysis, recommendation, cached]
      properties:
        id:
          type: ["integer", "null"]
        winScoreA:
          type: number
          description: Win score for Team A (0-100)
        winScoreB:
          type: number
          description: Win score for Team B (0-100)
        analysis:
          type: string
          description: AI-generated analysis of the trade
        recommendation:
          type: string
          description: Accept, Decline, or Neutral
        teamAName:
          type: ["string", "null"]
        teamBName:
          type: ["string", "null"]
        cached:
          type: boolean
        evaluatedAt:
          type: ["string", "null"]
          format: date-time

    LeagueSyncStatus:
      type: object
      required: [id, name, sport, autoSyncEnabled]
      properties:
        id:
          type: integer
        name:
          type: string
        sport:
          type: string
        autoSyncEnabled:
          type: boolean
        lastAutoSyncAt:
          type: ["string", "null"]
          format: date-time
        lastAutoSyncError:
          type: ["string", "null"]

    SyncStatus:
      type: object
      required: [hasCredentials, leagues]
      properties:
        hasCredentials:
          type: boolean
        leagues:
          type: array
          items:
            $ref: "#/components/schemas/LeagueSyncStatus"

    SyncRefreshResult:
      type: object
      required: [success, leaguesSynced, playersSynced, errors]
      properties:
        success:
          type: boolean
        leaguesSynced:
          type: integer
        playersSynced:
          type: integer
        errors:
          type: array
          items:
            type: string
        lastSyncAt:
          type: ["string", "null"]
          format: date-time

    DeleteResult:
      type: object
      required: [success, deleted]
      properties:
        success:
          type: boolean
        deleted:
          type: integer

    InsightTip:
      type: object
      required: [type, priority, message]
      properties:
        type:
          type: string
          description: "trade, waiver, lineup, or general"
        priority:
          type: string
          description: "high, medium, or low"
        player:
          type: ["string", "null"]
        message:
          type: string

    TeamInsightsResult:
      type: object
      required: [teamId, teamName, insights, tips, cached]
      properties:
        teamId:
          type: integer
        teamName:
          type: string
        insights:
          type: string
        tips:
          type: array
          items:
            $ref: "#/components/schemas/InsightTip"
        standingsRank:
          type: ["integer", "null"]
        totalTeams:
          type: ["integer", "null"]
        cached:
          type: boolean

    SuccessResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string

    DashboardSummary:
      type: object
      required: [leagueCount, teamCount, playerCount, tradeEvalCount]
      properties:
        leagueCount:
          type: integer
        teamCount:
          type: integer
        playerCount:
          type: integer
        tradeEvalCount:
          type: integer
        lastSyncAt:
          type: ["string", "null"]
          format: date-time
        topPlayers:
          type: array
          items:
            $ref: "#/components/schemas/Player"