# lib/api-zod/src/generated/types/espnPublicStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fantasy Helper API
 * OpenAPI spec version: 0.1.0
 */
import type { EspnPublicStatsStatLine } from './espnPublicStatsStatLine';

export interface EspnPublicStats {
  gamesPlayed: number;
  statLine: EspnPublicStatsStatLine;
  /** @nullable */
  injuryDescription?: string | null;
  /** @nullable */
  recentHeadline?: string | null;
  fetchedAt: Date;
}