# lib/api-zod/src/generated/types/player.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fantasy Helper API
 * OpenAPI spec version: 0.1.0
 */
import type { EspnPublicStats } from './espnPublicStats';

export interface Player {
  id: number;
  teamId: number;
  espnPlayerId: string;
  fullName: string;
  position: string;
  proTeam: string;
  /** @nullable */
  projectedPoints?: number | null;
  /** @nullable */
  avgPoints?: number | null;
  /** @nullable */
  totalPoints?: number | null;
  /** @nullable */
  injuryStatus?: string | null;
  espnPublicStats?: EspnPublicStats | null;
}